fix home page and singlepost

+++ src/components/Post.js
@@ -45,11 +45,14 @@ import '@fontsource/roboto/300.css';
 import '@fontsource/roboto/400.css';
 import '@fontsource/roboto/500.css';
 import '@fontsource/roboto/700.css';
+import Swal from "sweetalert2";
+import IconButton from "@mui/material/IconButton";
 
 const StyledPaper = styled(Typography)(({ theme }) => ({
   ...theme.typography.body2,
   maxWidth: 400,
 }));
+
 export default function Post({ post }) {
   // console.log(post.isLiked);
   const login = JSON.parse(localStorage.getItem("login"));
@@ -77,8 +80,20 @@ export default function Post({ post }) {
     setOpen(false);
   };
   const handleDeletePost = async (id) => {
-    await dispatch(deletePost(id));
-    handleClose();
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deletePost(id));
+      }
+    })
+    
   };
 
   const handleLike = async (isLiked) => {
@@ -89,9 +104,6 @@ export default function Post({ post }) {
         setLikeCount(likeCount - 1);
         await unlikePost(post._id);
     }
-  }
-  const checkAccessModified = async () => {
-    
   }
   return (
     <>
@@ -106,13 +118,12 @@ export default function Post({ post }) {
           <Avatar
             src={`${post.avatar}`}
             sx={{
-              border: "1px solid #ddd",
-              borderRadius: "4px",
-              padding: "5px",
+              borderRadius: "10",
               height: "150px",
               width: "200px",
             }}
             variant="rounded"
+            style={{ borderRadius: 10 }}
           />
         </TableCell>
         <TableCell align="left" width="800px">
@@ -155,8 +166,9 @@ export default function Post({ post }) {
             />
         </TableCell>
         <TableCell align="center">
-          <SmsIcon fontSize="small" />
-          <br />
+          <IconButton >
+              <SmsIcon fontSize="small" />
+            </IconButton>
           {post.comment}
         </TableCell>
         {post.author._id === userId && (
@@ -178,7 +190,7 @@ export default function Post({ post }) {
               color="warning"
               aria-label="delete"
               size="small"
-              onClick={() => handleClickOpen(post._id)}
+              onClick={() => handleDeletePost(post._id)}
             >
               <DeleteIcon />
             </Fab>
@@ -186,25 +198,6 @@ export default function Post({ post }) {
         )}
       </TableRow>
     </Container>
-      <Dialog
-        open={open}
-        onClose={handleClose}
-        aria-labelledby="alert-dialog-title"
-        aria-describedby="alert-dialog-description"
-      >
-        <DialogTitle id="alert-dialog-title">
-          {"Are you sure to delete the post?"}
-        </DialogTitle>
-        <DialogContent>
-          <DialogContentText id="alert-dialog-description"></DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleClose}>Disagree</Button>
-          <Button onClick={() => handleDeletePost(postId)} autoFocus>
-            Agree
-          </Button>
-        </DialogActions>
-      </Dialog>
     </>
   );
 }
